requirements.txt
Flask
botbuilder-core
botbuilder-schema
openai
azure-identity
azure-keyvault-secrets
